Make "go test" work on macOS
Abstract sockets, procfs and unix.Ucred are only available on Linux.
